+++ .atomist/tests/project/TypeScriptEditorSteps.ts
@@ -1,12 +1,15 @@
 import { Project } from "@atomist/rug/model/Project";
-import { Given, ProjectScenarioWorld, Then, When } from "@atomist/rug/test/project/Core";
+import {
+    Given, ProjectScenarioWorld, Then, When,
+} from "@atomist/rug/test/project/Core";
 
-When("the TypeScriptEditor is run", (p, world) => {
+When("the TypeScriptEditor is run", (p: Project, world) => {
     const w = world as ProjectScenarioWorld;
     const editor = w.editor("TypeScriptEditor");
     w.editWith(editor, { inputParameter: "the inputParameter value" });
 });
 
-Then("the hello file says hello", (p, world) => {
+Then("the hello file says hello", (p: Project, world) => {
+    const w = world as ProjectScenarioWorld;
     return p.fileContains("hello.txt", "Hello, World!");
 });