+++ .atomist/tests/project/TypeScriptGeneratorSteps.ts
@@ -1,14 +1,15 @@
 import { Project } from "@atomist/rug/model/Project";
-import { Given, ProjectScenarioWorld, Then, When } from "@atomist/rug/test/project/Core";
+import {
+    Given, ProjectScenarioWorld, Then, When,
+} from "@atomist/rug/test/project/Core";
 
-When("TypeScriptGenerator for TypeScriptGenerator should create a new project based on this archive", (p, world) => {
-    const psworld = world as ProjectScenarioWorld;
-    const generator = psworld.generator("TypeScriptGenerator");
-
-    psworld.generateWith(generator, "new-test-project", {});
+When("TypeScriptGenerator is run", (p: Project, world) => {
+    const w = world as ProjectScenarioWorld;
+    const generator = w.generator("TypeScriptGenerator");
+    w.generateWith(generator, "new-test-project", {});
 });
 
-Then("fileExists README md for TypeScriptGenerator should create a new project based on this archive", (p, world) => {
-
+Then("the README file exists", (p: Project, world) => {
+    const w = world as ProjectScenarioWorld;
     return p.fileExists("README.md");
 });